+++ docs/components/ApiView.vue
@@ -1,13 +1,15 @@
 <template>
     <section>
         <template v-for="component in data">
-            <h3 v-if="component.title" class="subtitle">{{ component.title }}</h3>
+            <h3 v-if="component.title" class="subtitle" :key="component.title">
+                {{ component.title }}
+            </h3>
             <b-tabs :key="component.title">
                 <b-tab-item v-if="component.props" label="Properties">
                     <b-table :data="component.props" default-sort="name">
                         <template slot-scope="props">
                             <b-table-column label="Name">
-                                <span v-html="props.row.name"></span>
+                                <span v-html="props.row.name" style="white-space: nowrap"></span>
                             </b-table-column>
                             <b-table-column label="Description">
                                 <span v-html="props.row.description"></span>
@@ -29,7 +31,7 @@
                     <b-table :data="component.slots" default-sort="name">
                         <template slot-scope="props">
                             <b-table-column label="Slot name">
-                                <span v-html="props.row.name"></span>
+                                <span v-html="props.row.name" style="white-space: nowrap"></span>
                             </b-table-column>
                             <b-table-column label="Description">
                                 <span v-html="props.row.description"></span>
@@ -45,7 +47,7 @@
                     <b-table :data="component.events" default-sort="name">
                         <template slot-scope="props">
                             <b-table-column label="Name">
-                                <span v-html="props.row.name"></span>
+                                <span v-html="props.row.name" style="white-space: nowrap"></span>
                             </b-table-column>
                             <b-table-column label="Description">
                                 <span v-html="props.row.description"></span>
@@ -61,7 +63,7 @@
                     <b-table :data="component.methods" default-sort="name">
                         <template slot-scope="props">
                             <b-table-column label="Name">
-                                <span v-html="props.row.name"></span>
+                                <span v-html="props.row.name" style="white-space: nowrap"></span>
                             </b-table-column>
                             <b-table-column label="Description">
                                 <span v-html="props.row.description"></span>