+++ docs/examples/icon/ExFa.vue
@@ -3,6 +3,24 @@
         <div class="block">
             <b-icon
                 pack="fa"
+                icon="user"
+                size="is-small">
+            </b-icon>
+            <b-icon
+                pack="fa"
+                icon="home"
+                size="is-small">
+            </b-icon>
+            <b-icon
+                pack="fa"
+                icon="dashboard"
+                size="is-small">
+            </b-icon>
+        </div>
+
+        <div class="block">
+            <b-icon
+                pack="fa"
                 icon="user">
             </b-icon>
             <b-icon
@@ -62,6 +80,15 @@
         <button class="button is-warning">
             <b-icon pack="fa" icon="check"></b-icon>
             <span>Finish</span>
+        </button>
+
+        <button class="button is-warning">
+            <b-icon
+                pack="fa"
+                icon="refresh"
+                custom-class="fa-spin">
+            </b-icon>
+            <span>Refresh</span>
         </button>
     </section>
 </template>