+++ docs/examples/icon/ExMdi.vue
@@ -1,14 +1,29 @@
 <template>
     <section>
         <div class="block">
-            <b-icon icon="person"></b-icon>
+            <b-icon
+                icon="account"
+                size="is-small">
+            </b-icon>
+            <b-icon
+                icon="home"
+                size="is-small">
+            </b-icon>
+            <b-icon
+                icon="view-dashboard"
+                size="is-small">
+            </b-icon>
+        </div>
+
+        <div class="block">
+            <b-icon icon="account"></b-icon>
             <b-icon icon="home"></b-icon>
-            <b-icon icon="dashboard"></b-icon>
+            <b-icon icon="view-dashboard"></b-icon>
         </div>
 
         <div class="block">
             <b-icon
-                icon="person"
+                icon="account"
                 size="is-medium">
             </b-icon>
             <b-icon
@@ -16,14 +31,14 @@
                 size="is-medium">
             </b-icon>
             <b-icon
-                icon="dashboard"
+                icon="view-dashboard"
                 size="is-medium">
             </b-icon>
         </div>
 
         <div class="block">
             <b-icon
-                icon="person"
+                icon="account"
                 size="is-large"
                 type="is-success">
             </b-icon>
@@ -33,19 +48,19 @@
                 type="is-info">
             </b-icon>
             <b-icon
-                icon="dashboard"
+                icon="view-dashboard"
                 size="is-large"
                 type="is-primary">
             </b-icon>
         </div>
 
         <button class="button is-dark">
-            <b-icon icon="done"></b-icon>
+            <b-icon icon="check"></b-icon>
             <span>Finish</span>
         </button>
 
         <button class="button is-warning">
-            <b-icon icon="done"></b-icon>
+            <b-icon icon="check"></b-icon>
             <span>Finish</span>
         </button>
     </section>