+++ docs/examples/select/ExIcons.vue
@@ -2,14 +2,14 @@
     <section>
         <h3 class="subtitle">With Material Design Icons</h3>
         <b-field>
-            <b-select placeholder="Country" icon="public">
+            <b-select placeholder="Country" icon="earth">
                 <option value="1">Option 1</option>
                 <option value="2">Option 2</option>
             </b-select>
         </b-field>
 
         <b-field>
-            <b-select placeholder="Client" icon="person">
+            <b-select placeholder="Client" icon="account">
                 <option value="1">Option 1</option>
                 <option value="2">Option 2</option>
             </b-select>