+++ docs/examples/upload/ExSimple.vue
@@ -2,7 +2,7 @@
     <b-field>
         <b-upload v-model="files">
             <a class="button is-primary">
-                <b-icon icon="file_upload"></b-icon>
+                <b-icon icon="upload"></b-icon>
                 <span>Click to upload</span>
             </a>
         </b-upload>