+++ src/components/panel/Panel.vue
@@ -5,15 +5,15 @@
             @click="toggle">
             <span v-if="header" v-html="header"></span>
             <slot v-else name="header"></slot>
-            <b-icon class="is-pulled-right" 
-                    v-if="collapsible" 
-                    both 
-                    :icon="isOpen ? 'arrow_drop_up' : 'arrow_drop_down'">
+            <b-icon class="is-pulled-right"
+                    v-if="collapsible"
+                    both
+                    :icon="isOpen ? 'menu-up' : 'menu-down'">
             </b-icon>
         </div>
         <transition :name="animation">
-            <div class="panel-content" 
-                 :class="{'panel-block' : !hasCustomTemplate}" 
+            <div class="panel-content"
+                 :class="{'panel-block' : !hasCustomTemplate}"
                  v-show="isOpen">
                 <div v-if="content" v-html="content"></div>
                 <slot v-else></slot>